requirements.txt
fastapi
os
shutil
groq
uvicorn
python-dotenv